mbti_analyzer/requirements.txt
# MBTI T/F 성향 분석기 의존성 패키지들

# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터 검증 및 모델링
pydantic==2.5.0
pydantic-settings==2.1.0

# AI 및 머신러닝
transformers==4.36.0
torch==2.1.1
google-generativeai==0.3.2

# HTTP 클라이언트
requests==2.31.0

# 환경변수 관리
python-dotenv==1.0.0

# 음성 처리 (선택적)
# speechrecognition==3.10.0
# pyaudio==0.2.11

# 개발 도구 (선택적)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0 